Refactor player messaging responsibilities to PlayerManager
